+++ packages/core/pluginutils.d.ts
@@ -7,3 +7,8 @@ export function hash(str: string): string
  * Dedent a string to look pretty
  */
 export function dedent(str: string): string
+
+/**
+ * Escape a string to be used as an attribute value
+ */
+export function escapeAttr(value: string): string