+++ packages/core/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.3 (2022-09-06)
 
 - Fix non-NodeNext type resolution
+- Fix `pluginutils` type
 
 ## 0.1.2 (2022-09-03)